% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whisker.path}
\alias{whisker.path}
\title{render.path}
\usage{
whisker.path(..., data = parent.frame(), fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character; defining a path}

\item{data}{named list or environment with variables that will be used during
rendering}

\item{fsep}{character; the path separator to use}
}
\description{
Combines \code{\link[base]{file.path}} except arguments are run through
\code{\link[whisker]{whisker.render}} first,
}
\note{
Whisker variables cannot contain '.' by specification, therefore the
  template cannot use names with '.'
}
\examples{
dir <- "~/tmp"
  file_name <- "myfile"
  whisker.path( "{{dir}}", "{{file_name}}.txt" )
}

