% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/whisker.render.dir.R
\name{whisker.render.dir}
\alias{whisker.render.dir}
\title{Render all the templates in a directory}
\usage{
whisker.render.dir(path, out = ".", data = parent.frame(), ...)
}
\arguments{
\item{path}{character; path to single directory containing templates}

\item{out}{character; path to single directory where rendered files are
written.}

\item{data}{named list or environment with variables that will be used
during rendering}

\item{...}{additional arguments}
}
\description{
Render all the templates in a directory
}

