% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker_check.R
\name{whisker_check}
\alias{whisker_check}
\title{whisker_check}
\usage{
whisker_check(template, data)
}
\arguments{
\item{template}{character; mustache template}

\item{data}{named list or environment used for template construction

\code{whisker_check} validates that all the values for a template are present.}
}
\value{
Returns \code{TRUE} if all template names are available. False otherwise.
}
\description{
Check a template for all variables
}
\examples{

  tmpl <- "Hello {{Who}}"
  whisker_check( tmpl, list(Who = "World" ))     # TRUE
  whisker_check( tmpl, list(Person = "World" ))  # FALSE 

}
\seealso{
\code{\link{whisker_names}} \cr
}
