% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whisker_extract_all}
\alias{whisker_extract_all}
\title{whisker_extract_all}
\usage{
whisker_extract_all(template)
}
\arguments{
\item{template}{character of template text

 Provides a stringr type function for extracting the names for the
 substituted variable. This is useful if you want to do some checking or
 other manipulation of the template prior to rendering.}
}
\value{
chracter vector of names found in the template
}
\description{
Extract varaibles from a whisker template
}
\examples{
template <- "Dear {{name}}, I have not loved you since {{date}}"

   whisker_extract_all(template)
}
\seealso{
\code{\link[whisker]{whisker.render}}
   \code{\link[stringr]{str_extract_all}} on which this function relies
}

