% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/whisker_get_all.R
\name{whisker_get_all}
\alias{whisker_get_all}
\title{Get template values from the call stack}
\usage{
whisker_get_all(template, ...)
}
\arguments{
\item{template}{character; a whisker template}

\item{...}{additional arguments passed to \code{\link[base]{get}}

\code{whisker_get_all} searches the call stack to find the variables in a
whisker template.}
}
\value{
list containing the names and values from the template. If a value is not
  found, \code{NA} is returned.
}
\description{
Introspect template and get values from their first appearance on the call
stack.
}
\examples{
a=1; z=26

  tmpl = "A is rank {{a}}; z is rank {{z}}; {{w}}"
  whisker_get_all( tmpl )

  # In function:
  f <- function( template ) whisker_get_all( template )
  f(tmpl)
}
\seealso{
\code{\link[base]{get}} \cr
  \code{\link{whisker_names}} \cr
}

