% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/whisker_names.R
\name{whisker_names}
\alias{whisker_names}
\title{whisker_names}
\usage{
whisker_names(template)
}
\arguments{
\item{template}{character of template text

 Provides a stringr type function for extracting the names for the
 substituted variable. This is useful if you want to do some checking or
 other manipulation of the template prior to rendering.}
}
\value{
chracter vector of names found in the template
}
\description{
Extract varaibles names from a whisker template
}
\examples{
template <- "Dear {{name}}, I have not loved you since {{date}}"

   whisker_names(template)
}
\seealso{
\code{\link[whisker]{whisker.render}}
   \code{\link[stringi]{stri_extract_all_regex}} on which this function relies
}

