% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker_path.R
\name{whisker_path}
\alias{whisker.path}
\alias{whisker_path}
\title{whisker_path}
\usage{
whisker_path(..., data = parent.frame(), fsep = .Platform$file.sep)

whisker.path(..., data = parent.frame(), fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character; defining a path}

\item{data}{named list or environment with variables that will be used during
rendering}

\item{fsep}{character; the path separator to use}
}
\description{
Combines \code{\link[base]{file.path}} except arguments are run through 
\code{\link[whisker]{whisker.render}} first,
}
\details{
\code{whisker.path} is an alias for \code{whisker_path.}
}
\note{
Whisker variables cannot contain '.' by specification, therefore the 
  template cannot use names with '.'
}
\examples{
  dir <- "~/tmp"
  file_name <- "myfile"
  whisker_path( "{{dir}}", "{{file_name}}.txt" )
  
}
\seealso{
\code{\link[base]{file.path}} \cr
  \code{\link[whisker]{whisker.render}}
}

