% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker_render.R
\name{whisker_render}
\alias{whisker_render}
\title{whisker_render}
\usage{
whisker_render(template, data = parent.frame(), partials = list(),
  debug = FALSE)
}
\description{
An alias for whisker.render
}
\seealso{
\code{\link[whisker]{whisker.render}}
}

