% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whisker.render.R
\name{whisker.render...}
\alias{whisker.render...}
\title{whisker.render... - take variables from the ... arguments}
\usage{
whisker.render...(...)
}
\arguments{
\item{...}{values to be subsituted}
}
\value{
character table centeres
}
\description{
whisker.render... - take variables from the ... arguments
}

