% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{whisker.tools}
\alias{whisker.tools}
\alias{whisker.tools-package}
\title{whisker.tools}
\description{
Methods that extend the whisker package
}
\references{
\url{http://cran.r-project.org/web/packages/whisker/index.html}
}

