% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whisker.tools-package.R
\docType{package}
\name{whisker.tools}
\alias{whisker.tools}
\alias{whisker.tools-package}
\title{whisker.tools}
\description{
Methods that extend the whisker package
}
\references{
\url{http://cran.r-project.org/web/packages/whisker/index.html}
}

