% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker_names.R
\name{whisker_names}
\alias{whisker_names}
\title{whisker_names}
\usage{
whisker_names(template)
}
\arguments{
\item{template}{character of template text

Provides a stringr type function for extracting the names for the
substituted variable. This is useful if you want to do some checking or
other manipulation of the template prior to rendering.}
}
\value{
chracter vector of names found in the template
}
\description{
Extract varaibles names from a whisker template
}
\examples{
  template <- "Dear {{name}}, I have not loved you since {{date}}"

  whisker_names(template)


}
\references{
See for example, \url{http://stackoverflow.com/a/16596015/1504321}
}
\seealso{
\code{\link{whisker_render}} \cr
  \code{\link[stringi]{stri_extract_all_regex}} on which this function relies
}

