% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker_render.R
\name{whisker_render}
\alias{whisker_render}
\title{whisker_render}
\usage{
whisker_render(
  template,
  data = parent.frame(),
  partials = list(),
  debug = FALSE,
  strict = TRUE
)
}
\arguments{
\item{template}{character with template text}

\item{data}{named list or environment with variables that will be used during rendering}

\item{partials}{named list with partial templates, will be used during template construction}

\item{debug}{Used for debugging purposes, likely to disappear}
}
\description{
An alias for whisker.render
}
\seealso{
\code{\link[whisker]{whisker.render}}
}
\author{
Edwin de Jonge, \url{https://github.com/edwindj/whisker}
}
